:sparkles: Se realiza la configuración del consumo de cada uno de los path definidos desde el backend para mostrar un mensaje desde la aplicación.
