+++ src/components/pages/Login.jsx
@@ -1,9 +1,31 @@
+import React, { useEffect } from "react";
 import "../../css/FormularioLogin.css";
 import Form from "react-bootstrap/Form";
 
-export default function login() {
+export default function Login() {
+	const [data, setData] = React.useState("Esperando data...");
+
+	const getData = async () => {
+		try {
+			fetch("http://localhost:3000/login")
+				.then((res) => res.json())
+				.then((data) => setData(data.message));
+		} catch (error) {
+			console.error(error);
+		}
+	};
+
+	useEffect(() => {
+		(async () => {
+			await getData();
+		})();
+	}, []);
+
 	return (
 		<div className="container">
+			
+			<h4>{!data ? "Loading..." : data}</h4>
+
 			<Form className="formulario-login">
 				<h4 className="mt-5">Iniciar sesión</h4>
 				<Form.Group className="mb-3" controlId="formBasicEmail">
